chore: move tabbables to libs dir
